+++ packages/notifications/src/components/NotificationMenu/styled/index.ts
@@ -1,36 +1,41 @@
 import styled from 'styled-components';
 import { LoadingShape, Spinner } from 'react-rainbow-components';
+import { CloseCircleFilled } from '@rainbow-modules/icons';
 import { NotificationProps, StatusBadgeProps } from '../types';
 
 export const StyledSpinner = styled(Spinner)`
     position: relative;
+    z-index: 1;
+    top: unset;
+    left: unset;
+    transform: unset;
 `;
 
 export const NotificationMenuContainer = styled.div`
     position: relative;
-    display: inline-block;
+    display: block;
 `;
 
 export const DropdownContainer = styled.div.attrs((props) => props.theme.rainbow)`
     background-color: ${(props) => props.palette.background.main};
     color: ${(props) => props.palette.text.main};
     box-shadow: ${(props) => props.shadows.shadow_2};
-    width: 24rem;
+    width: 450px;
     border-radius: 0.875rem;
 `;
 
-export const Header = styled.div.attrs((props) => props.theme.rainbow)`
+export const Header = styled.header.attrs((props) => props.theme.rainbow)`
     display: flex;
     justify-content: space-between;
     align-items: center;
-    padding: 0.6rem;
+    padding: 0.85rem 1rem;
     border-bottom: solid 1px ${(props) => props.palette.border.disabled};
-    font-size: 1.25rem;
+    font-size: 18px;
     font-weight: bold;
 `;
 
-export const Footer = styled.div.attrs((props) => props.theme.rainbow)`
-    padding: 0.2rem;
+export const Footer = styled.footer.attrs((props) => props.theme.rainbow)`
+    padding: 0.5rem;
     border-top: solid 1px ${(props) => props.palette.border.disabled};
 `;
 
@@ -41,14 +46,18 @@ export const Notifications = styled.ul`
 export const NotificationContainer = styled.li.attrs((props) => props.theme.rainbow)`
     position: relative;
     display: flex;
-    padding: 0.6rem 0.6rem 0.6rem 1rem;
+    padding: 20px 12px 14px 16px;
     background-color: ${(props) => props.palette.background.secondary};
     cursor: pointer;
 
-    ${(props) => props.isLoading && `background-color: ${props.palette.background.main};`}
+    :hover {
+        background-color: ${(props) => props.palette.action.hover};
+    }
+
+    ${(props) => props.isLoading && `background-color: transparent`}
 
     :not(:last-of-type) {
-        border-bottom: solid 1px ${(props) => props.palette.border.divider};
+        border-bottom: solid 1px ${(props) => props.palette.border.disabled};
         ${(props) => props.isLoading && `border-bottom: none;`}
     }
 
@@ -58,12 +67,13 @@ export const NotificationContainer = styled.li.attrs((props) => props.theme.rain
             !props.hasIcon &&
             `
             position: absolute;
-            top: 1rem;
-            left: 6px;
+            top: 1.75rem;
+            left: 10px;
             content: '';
             display: block;
             width: 8px;
             height: 8px;
+            box-shadow: 0 0 0 1px ${props.palette.border.disabled};
             border-radius: 50%;
             background-color: ${props.palette.brand.main};
             `}
@@ -75,7 +85,7 @@ export const Content = styled.div.attrs((props) => props.theme.rainbow)`
     display: flex;
     flex-direction: column;
     flex-grow: 1;
-    margin-left: 6px;
+    margin-left: 12px;
 `;
 
 export const TitleContainer = styled.div<{ isLoading?: boolean }>`
@@ -83,7 +93,6 @@ export const TitleContainer = styled.div<{ isLoading?: boolean }>`
     display: flex;
     flex-direction: row;
     justify-content: space-between;
-    margin-bottom: 0.2rem;
 
     ${(props) => props.isLoading && `margin-bottom: 0.6rem;`}
 `;
@@ -93,10 +102,10 @@ export const Title = styled.div.attrs((props) => props.theme.rainbow)<Notificati
     overflow: hidden;
     text-overflow: ellipsis;
     color: ${(props) => props.palette.text.main};
-    font-weight: bold;
-    font-size: 1rem;
+    font-size: 16px;
+    margin-bottom: 4px;
 
-    ${(props) => props.unread && `font-weight: bold;`}
+    ${(props) => props.unread && `font-family: Lato Bold, Helvetica, sans-serif;`}
 `;
 
 export const CreatedAt = styled.div.attrs((props) => props.theme.rainbow)<NotificationProps>`
@@ -105,24 +114,31 @@ export const CreatedAt = styled.div.attrs((props) => props.theme.rainbow)<Notifi
     text-overflow: ellipsis;
     max-width: 35%;
     color: ${(props) => props.palette.text.label};
-    font-size: 0.875rem;
+    font-size: 14px;
+    opacity: 0.9;
 
     ${(props) =>
         props.unread &&
         `
-        font-weight: bold;
+        font-family: Lato Bold, Helvetica, sans-serif;
         color: ${props.palette.text.main};
+        opacity: 1;
         `}
 `;
 
 export const Description = styled.div.attrs((props) => props.theme.rainbow)<NotificationProps>`
     color: ${(props) => props.palette.text.label};
-    font-size: 0.875rem;
+    font-size: 14px;
+    margin-bottom: 4px;
+    opacity: 0.9;
+    line-height: 1.5;
+
     ${(props) =>
         props.unread &&
         `
-        font-weight: bold;
+        font-family: Lato Bold, Helvetica, sans-serif;
         color: ${props.palette.text.main};
+        opacity: 1;
         `}
 `;
 
@@ -130,45 +146,42 @@ export const StatusBadge = styled.span.attrs((props) => props.theme.rainbow)<Sta
     display: flex;
     align-items: center;
     text-transform: uppercase;
-    color: ${(props) => props.palette.success.main};
-    background-color: ${(props) => props.palette.background.secondary};
+    color: ${(props) => props.palette.success.dark};
+    background-color: transparent;
     padding: 0;
     margin-top: 4px;
+    font-size: 14px;
 
     ${(props) => props.status === 'error' && `color: ${props.palette.error.main}`}
     ${(props) => props.status === 'warning' && `color: ${props.palette.warning.main}`}
     ${(props) => props.status === 'inProgress' && `color: ${props.palette.brand.main}`}
+    ${(props) => props.status === 'info' && `color: ${props.palette.brand.main}`}
 `;
 
-export const StatusIconContainer = styled.div.attrs((props) => props.theme.rainbow)`
-    display: flex;
-    align-items: center;
-    justify-content: center;
+export const StatusIcon = styled(CloseCircleFilled).attrs((props) => props.theme.rainbow)`
     width: 16px;
     height: 16px;
-    padding: 2px;
-    border-radius: 50%;
     margin-right: 0.4rem;
-    background-color: ${(props) => props.palette.brand.main};
+
     ${(props) =>
         props.status === 'success' &&
         `
-        background-color: ${props.palette.success.main};
+        color: ${props.palette.success.dark};
     `}
     ${(props) =>
         props.status === 'warning' &&
         `
-        background-color: ${props.palette.warning.main};
+        color: ${props.palette.warning.main};
     `}
     ${(props) =>
         props.status === 'error' &&
         `
-        background-color: ${props.palette.error.main};
+        color: ${props.palette.error.main};
     `}
     ${(props) =>
-        props.status === 'inProgress' &&
+        props.status === 'info' &&
         `
-        background-color: ${props.palette.background.main};
+        color: ${props.palette.brand.main};
     `}
 `;